Apply fix for SBST to SBSTBase. Also fix some bugs idiosyncratic to SBST: namely, ensure values never get rounded to zero.

+++ src/main/java/uk/ac/cam/cl/arg58/mphil/compression/IntegerMass.java
@@ -6,5 +6,5 @@ import uk.ac.cam.eng.ml.tcs27.compression.Mass;
 public interface IntegerMass extends Mass<Integer> {
   /* start and end are inclusive */
   public double massBetween(Integer start, Integer end);
-  public int discreteMassBetween(Integer start, Integer end);
+  public long discreteMassBetween(Integer start, Integer end);
 }
